+++ CHANGELOG.md
@@ -4,6 +4,7 @@ FEATURES
 
 * **New Analyzer:** `schemavalidatefunc`: returns `SchemaValidateFunc` declarations
 * **New Check:** `R005`: check for `ResourceData.HasChange()` calls that can be combined into one `HasChanges()` call
+* **New Check:** `S021`: check for `Schema` that should omit `ComputedWhen`
 * **New Check:** `V001`: check for custom `SchemaValidateFunc` that implement `validation.StringMatch()` or `validation.StringDoesNotMatch()`
 
 ENHANCEMENTS